make it so we do not hemmorage memory, mostly through not freeing textures and surfaces. also added a sheilded spawn area. It's kinda a cool game now. Who woulda guessed.

+++ game/src/EnemyObject.cpp
@@ -4,10 +4,10 @@ EnemyObject::EnemyObject(int x, int y, int xVelocity, int yVelocity,
                          EnemyInputComponent *i, EnemyGraphicsComponent *g,
                          EnemyPhysicsComponent *p, int entityNum) {
     location = {x, y, 0, 0};
-    this->xFloat = x;
-    this->yFloat = y;
-    this->xVelocity = xVelocity;
-    this->yVelocity = yVelocity;
+    this->xFloat = (float)x;
+    this->yFloat = (float)y;
+    this->xVelocity = (float)xVelocity;
+    this->yVelocity = (float)yVelocity;
     this->input = i;
     this->graphics = g;
     this->physics = p;
